refactored validation completely to make validation checks easier in the future
