+++ langium/src/language/services/jpipe-hover-provider.ts
@@ -1,9 +1,8 @@
 import { type AstNode, type MaybePromise, } from 'langium';
 import { AstNodeHoverProvider } from 'langium/lsp';
 import { Hover } from 'vscode-languageserver';
-import { isClassType, isInstructionType } from './jpipe-completion-provider.js';
+import { isClassType, isInstructionType } from './validation/main-validation.js';
 import { isVariable } from '../generated/ast.js';
-
 
 //provides hover for variables and class types
 export class JpipeHoverProvider extends AstNodeHoverProvider{
